Move functions to an importable library
Previously the code was difficult to test. To make it testable, the main
function now returns a string rather than printing, and it's name
was changed. Additionally, all functions have been moved to a lib
that can be imported by both the cli and the tests. These changes
can be compared using 'git diff -C'