Add ability to load all datasets within a prefix
Previously, when using p2h, each dataset had to be manually added -
that is, p2h would have to be called on each one individually. Usually
this is done in a cron job.

With this change, one can simply call p2h on a bucket/prefix, and
all valid datasets within that bucket/prefix will be loaded.

When using this option, not all the usual directives are available,
for example `alias`. However, success-only, use-last-versions
and exclude-regex will still be available,
and will be used when loading every dataset.

ref: bug 1301386